move the config file, move commonly used functions to common
