+++ rt1_input_pj7.py
@@ -4,19 +4,15 @@ os.environ['OPENBLAS_NUM_THREADS'] = '1'
 os.environ['MKL_NUM_THREADS'] = '1'
 os.environ['OMP_NUM_THREADS'] = '1'
 os.environ['NUMEXPR_NUM_THREADS'] = '1'
-import cloudpickle
 
 from ImageStack import create_imagestack_dataset
 from datetime import datetime
 import pandas as pd
 import numpy as np
-from configparser import ConfigParser
-import argparse
-import copy
 from rt1_processing_funcs_juhuu import inpdata_inc_average
 import random, string
 from scipy.signal import savgol_filter
-
+from common import chunkIt, parse_args, read_cfg, parallelfunc
 
 
 def read_stack(sig0_dir, plia_dir, block_size, cr_list, output_dir, ndvi_dir=None, orbit_direction=''):
@@ -52,24 +48,22 @@ def read_stack(sig0_dir, plia_dir, block_size, cr_list, output_dir, ndvi_dir=Non
     else:
         ndvi_stack = None
 
-
     # prepare timelist and filelist for sig0 reader
     files_sig0 = os.listdir(sig0_dir)
     filelist_sig0 = []
     times_sig0 = []
-    sig0_pattern = 'IWGRDH1VV%s' %orbit_direction
+    sig0_pattern = 'IWGRDH1VV%s' % orbit_direction
 
     for fil in files_sig0:
         if 'SIG' in fil and sig0_pattern in fil and fil.endswith('.tif'):
             filelist_sig0.append(os.path.join(sig0_dir, fil))
             times_sig0.append(datetime.strptime(fil[1:16], "%Y%m%d_%H%M%S"))
 
-
     # prepare timelist and filelist for plia reader
     files_plia = os.listdir(plia_dir)
     filelist_plia = []
     times_plia = []
-    plia_pattern = 'IWGRDH1--%s' %orbit_direction
+    plia_pattern = 'IWGRDH1--%s' % orbit_direction
 
     for fil in files_plia:
         if 'PLIA' in fil and plia_pattern in fil and fil.endswith('.tif'):
@@ -80,7 +74,6 @@ def read_stack(sig0_dir, plia_dir, block_size, cr_list, output_dir, ndvi_dir=Non
     if len(times_sig0) != len(times_plia):
         print("Warning! The number of sig0 and plia images is not equal")
 
-
     # check if there is any files in sig0 which the correspond plia doesn't exist
     for time in times_sig0:
         if time not in times_plia:
@@ -88,7 +81,6 @@ def read_stack(sig0_dir, plia_dir, block_size, cr_list, output_dir, ndvi_dir=Non
             del times_sig0[idx_sig0_no_plia]
             del filelist_sig0[idx_sig0_no_plia]
             print("Warning! The scene in this date ", time, " doesn't have the correspond plia. Removing...")
-
 
     # read sig and plia blocks
     for [c, r] in cr_list:
@@ -100,15 +92,14 @@ def read_stack(sig0_dir, plia_dir, block_size, cr_list, output_dir, ndvi_dir=Non
         # read ndvi to virtual stack
         if ndvi_dir:
             ndvi_stack = create_imagestack_dataset(name=random_name + 'ndvi', filelist=filelist_ndvi, times=times_ndvi,
-                                               nodata=-9999)
+                                                   nodata=-9999)
 
         # read sig0 to virtual stack
         sig_stack = create_imagestack_dataset(name=random_name + 'SIG', filelist=filelist_sig0, times=times_sig0,
-                                      nodata=-9999)
+                                              nodata=-9999)
         # read plia to virtual stack
         plia_stack = create_imagestack_dataset(name=random_name + 'LIA', filelist=filelist_plia, times=times_plia,
-                                       nodata=-9999)
-
+                                               nodata=-9999)
 
         sig0_block = sig_stack.read_ts(c * block_size, r * block_size, block_size, block_size)
         plia_block = plia_stack.read_ts(c * block_size, r * block_size, block_size, block_size)
@@ -179,8 +170,6 @@ def read_stack(sig0_dir, plia_dir, block_size, cr_list, output_dir, ndvi_dir=Non
             # sort by index
             df_ndvi.sort_index(inplace=True)
 
-
-
         # ------------------------ RQ's manipulation
         '''
         manual_dyn_df = pd.DataFrame(df.index.month.values.flatten(), df.index, columns=['VOD'])
@@ -210,7 +199,7 @@ def read_stack(sig0_dir, plia_dir, block_size, cr_list, output_dir, ndvi_dir=Non
 
         VOD_input = df_ndvi.resample('D').interpolate(method='nearest')
         # get a smooth curve
-        #VOD_input = VOD_input.rolling(window=60, center=True, min_periods=1).mean()
+        # VOD_input = VOD_input.rolling(window=60, center=True, min_periods=1).mean()
         VOD_input = VOD_input.clip_lower(0).apply(savgol_filter, window_length=61, polyorder=2).clip_lower(0)
         # reindex to input-dataset
         VOD_input = VOD_input.reindex(df.index.drop_duplicates()).dropna()
@@ -219,126 +208,21 @@ def read_stack(sig0_dir, plia_dir, block_size, cr_list, output_dir, ndvi_dir=Non
         # drop all measurements where no VOD estimates are available
         df = df.loc[VOD_input.dropna().index]
 
-        #manual_dyn_df = pd.DataFrame(df.index.month.values.flatten(), df.index, columns=['VOD'])
+        # manual_dyn_df = pd.DataFrame(df.index.month.values.flatten(), df.index, columns=['VOD'])
         defdict_i = {
-                    'bsf'   : [False, 0.01, None,  ([0.01], [.25])],
-                    'v'     : [False, 0.4, None, ([0.01], [.4])],
-                    'v2'    : [True, 1., None, ([0.5], [1.5])],
-                    'VOD'   : [False, VOD_input.values.flatten()],
-                    'SM'    : [True, 0.25,  'D',   ([0.05], [0.5])],
-                    #'VOD'   : [True, 0.25,'30D', ([0.01], [1.])],
-                    'frac'  : [True, 0.5, None,  ([0.01], [1.])],
-                    'omega' : [True, 0.3,  None,  ([0.05], [0.6])],
-                    }
+            'bsf': [False, 0.01, None, ([0.01], [.25])],
+            'v': [False, 0.4, None, ([0.01], [.4])],
+            'v2': [True, 1., None, ([0.5], [1.5])],
+            'VOD': [False, VOD_input.values.flatten()],
+            'SM': [True, 0.25, 'D', ([0.05], [0.5])],
+            # 'VOD'   : [True, 0.25,'30D', ([0.01], [1.])],
+            'frac': [True, 0.5, None, ([0.01], [1.])],
+            'omega': [True, 0.3, None, ([0.05], [0.6])],
+        }
 
         # TODO: pass ndvi df to out_dict
         out_dict = {'dataset': df, 'defdict': defdict_i, '_fnevals_input': None, 'c': c, 'r': r, 'outdir': output_dir}
         parallelfunc(out_dict)
-
-
-
-def chunkIt(seq, num):
-    '''
-    Chunk a list in to a approximately equal length
-    Parameters
-    ----------
-    seq: list
-    num: number of part
-
-    Returns
-    -------
-
-    '''
-    avg = len(seq) / float(num)
-    out = []
-    last = 0.0
-
-    while last < len(seq):
-        out.append(seq[int(last):int(last + avg)])
-        last += avg
-
-    return out
-
-
-def parse_args(args):
-    """
-    Parse command line parameters.
-
-    Parameters
-    ----------
-    args: list
-        command line arguments
-
-    Returns
-    -------
-    parser: parsed arguments
-    """
-    parser = argparse.ArgumentParser(
-        description="RT1 processing")
-
-    parser.add_argument(
-        "cfg_file", help="Path to config file for this processing step")
-
-    parser.add_argument("-arraynumber", "--arraynumber", metavar='arraynumber', type=str,
-                        help=("Number of VSC3 array (int)"))
-
-    parser.add_argument("-totalarraynumber", "--totalarraynumber", metavar='totalarraynumber', type=int,
-                        help=("Total number of VSC3 arrays (int)"))
-
-    return parser.parse_args(args)
-
-
-def read_cfg(cfg_file, include_default=True):
-    """
-    Parse a config file.
-
-    Parameters
-    ----------
-    cfg_file: string
-        filename of the config file
-    include_default: boolean, optional
-        If true then the DEFAULT section settings will
-        be included in all other sections.
-    """
-
-    config = ConfigParser()
-    config.optionxform = str
-    config.read(cfg_file)
-
-    ds = {}
-
-    for section in config.sections():
-
-        ds[section] = {}
-        for item, value in config.items(section):
-
-            if 'path' in item:
-                value = value.replace(' ', '')
-                path = value.split(',')
-                if path[0][0] == '.':
-                    # relative path
-                    value = os.path.join(os.path.split(cfg_file)[0], *path[0:])
-                elif path[0][0] == '/' or path[0][1] == ':':
-                    # absolute path in linux or windows
-                    value = os.path.sep.join(path)
-                else:
-                    print(section, item, ' got a blank value, set to  None')
-                    value = None
-
-            if item.startswith('kws'):
-                if item[4:] == 'custom_dtype':
-                    value = {item[4:]: eval(value)}
-                else:
-                    value = {item[4:]: value}
-                item = 'kws'
-
-            if include_default or item not in config.defaults().keys():
-                if item == 'kws':
-                    ds[section][item].update(value)
-                else:
-                    ds[section][item] = value
-
-    return ds
 
 
 def main(args, test_vsc_param=False):
@@ -425,7 +309,7 @@ def main(args, test_vsc_param=False):
         print('len_cr_list (px for this node)', len(list_to_process_node))
 
     else:
-        # print("load the stack...:", datetime.now())
+        # single thread processing
         if mp_threads in [0, 1]:
             read_stack(sig0_dir=sig0_dir,
                        plia_dir=plia_dir,
@@ -437,9 +321,8 @@ def main(args, test_vsc_param=False):
         else:
             # implement the multiprocessing here
             import multiprocessing as mp
-
             process_list = []
-            list_to_process_node_chunked = chunkIt(list_to_process_node, mp_threads*20)
+            list_to_process_node_chunked = chunkIt(list_to_process_node, mp_threads * 20)
             for cr_list in list_to_process_node_chunked:
                 process_dict = {}
                 process_dict['sig0_dir'] = sig0_dir
@@ -451,154 +334,57 @@ def main(args, test_vsc_param=False):
                 process_dict['orbit_direction'] = orbit_direction
                 process_list.append(process_dict)
 
-
             print("start the mp...:", datetime.now())
             print('processing ', len(process_list), 'sites...')
             pool = mp.Pool(mp_threads)
             pool.map(read_stack_mp, process_list)
             pool.close()
 
-            #import ipyparallel as ipp
-            #client = ipp.Client()
-            #dview = client[:]
-            #dview.use_cloudpickle()
-            #balanced_view = client.load_balanced_view()
-
-            #async_res = balanced_view.map_async(read_stack_mp, process_list)
-            #return async_res
 
 def read_stack_mp(process_dict):
-
     read_stack(sig0_dir=process_dict['sig0_dir'],
                plia_dir=process_dict['plia_dir'],
                block_size=process_dict['block_size'],
                cr_list=process_dict['cr_list'],
                output_dir=process_dict['output_dir'],
-               ndvi_dir = process_dict['ndvi_dir'],
-               orbit_direction = process_dict['orbit_direction'],
-               )
+               ndvi_dir=process_dict['ndvi_dir'],
+               orbit_direction=process_dict['orbit_direction'])
 
-# -------------------------------------------------
-
-
-def parallelfunc(import_dict):
-    import os
-    os.environ['OPENBLAS_NUM_THREADS'] = '1'
-    os.environ['MKL_NUM_THREADS'] = '1'
-    os.environ['OMP_NUM_THREADS'] = '1'
-    os.environ['NUMEXPR_NUM_THREADS'] = '1'
-
-    c = import_dict['c']
-    r = import_dict['r']
-    outdir = import_dict['outdir']
-    dataset = copy.deepcopy(import_dict['dataset'])
-    dataset['sig'] = 10 ** (dataset['sig'] / 10.)
-
-    defdict = import_dict['defdict']
-    _fnevals_input = import_dict['_fnevals_input']
-    print('processing site', c, r)
-
-    from rt1.rtfits import Fits
-    '''
-    def set_V_SRF(isopart, alpha, v1, v2, v_forward, v_bounceoff, VOD,
-                  omega, tt, s1, s2, aa, SM, **kwargs):
-        from rt1.volume import LinCombV, HenyeyGreenstein
-        from rt1.surface import HG_nadirnorm
-
-        V = LinCombV(Vchoices=
-                     [[isopart,
-                       HenyeyGreenstein(t=0.001, ncoefs=3)],
-                      [(1. - isopart) * (1. - alpha),
-                       HenyeyGreenstein(t=v_forward, ncoefs=8)],
-                      [(1. - isopart) * alpha,
-                       HenyeyGreenstein(t=-v_bounceoff, ncoefs=8, a=[1., 1., 1.])]
-                      ],
-                     tau=v1 + v2 * VOD,
-                     omega=omega)
-
-        SRF = HG_nadirnorm(t=tt, ncoefs=10, a=[aa, 1., 1.],
-                           NormBRDF=s1 + s2 * SM)
-        return V, SRF
-    '''
-    def set_V_SRF(frac, omega, SM, VOD, v, v2, **kwargs):
-
-        from rt1.volume import HenyeyGreenstein
-        from rt1.surface import LinCombSRF, HG_nadirnorm
-
-        SRFchoices = [
-                [frac, HG_nadirnorm(t=0.01, ncoefs=2, a=[-1.,1.,1.])],
-                [(1.-frac), HG_nadirnorm(t=0.6, ncoefs=10, a=[1., 1., 1.])]
-                ]
-        SRF = LinCombSRF(SRFchoices = SRFchoices, NormBRDF=SM)
-
-        V = HenyeyGreenstein(t=v, omega=omega, tau=v2*VOD, ncoefs=8)
-
-        return V, SRF
-
-
-
-    fit = Fits(sig0=True, dB=False, dataset=dataset,
-               set_V_SRF=set_V_SRF,
-               defdict=defdict)
-
-    fitset = {'int_Q': False,
-              '_fnevals_input': _fnevals_input,
-              # least_squares kwargs:
-              'verbose': 1,  # verbosity of least_squares
-              # 'verbosity' : 1, # verbosity of monofit
-              'ftol': 1.e-5,
-              'gtol': 1.e-5,
-              'xtol': 1.e-5,
-              'max_nfev': 100,
-              'method': 'trf',
-              'tr_solver': 'lsmr',
-              'x_scale': 'jac'
-              }
-
-    fit.performfit(**fitset)
-    fit.result[1].fn = 1
-
-    if import_dict['_fnevals_input'] is None:
-        import_dict['_fnevals_input'] = fit.result[1]._fnevals
-
-    with open(os.path.join(outdir, str(c) + '_' + str(r) + '.dump'), 'wb') as file:
-        cloudpickle.dump(fit, file)
-        # return fit
 
 if __name__ == '__main__':
     import sys
-    #sys.argv.append(r"E:\USERS\tle\rt1_input_pj7\src\rt1_input\config_pj7.ini")
-    #sys.argv.append("-totalarraynumber")
-    #sys.argv.append("1")
-    #sys.argv.append("-arraynumber")
-    #sys.argv.append("1")
 
+    # sys.argv.append(r"E:\USERS\tle\rt1_input_pj7\src\rt1_input\config_pj7.ini")
+    # sys.argv.append("-totalarraynumber")
+    # sys.argv.append("1")
+    # sys.argv.append("-arraynumber")
+    # sys.argv.append("1")
 
     print("Start", datetime.now())
     async_res = main(sys.argv[1:], test_vsc_param=False)
 
-    #print('gogogo')
+    # print('gogogo')
     # initialize a stdout0 array for comparison
 
-#    import time as tme
-#    vectorlen = np.vectorize(len)
-#
-#    stdout0 = np.array(async_res.stdout)
-#    while not async_res.ready():
-#        #ids = async_res.engine_id
-#        stdout1 = np.array(async_res.stdout)
-#        where = stdout1 != stdout0
-#        if np.any(where):
-#            oldlens = vectorlen(stdout0[where])
-#            changedstdouts = ['Kernel '
-#                              #+ str(ids[i]) + ':  '
-#                              + str(np.arange(len(where))[where][i]) + ':  '
-#                              + '\n'
-#                              + val[oldlength:]
-#                              for i, [val, oldlength] in enumerate(zip(stdout1[where], oldlens))]
-#
-#            sys.stdout.write('\r ' + ' '.join(changedstdouts))
-#            sys.stdout.flush()
-#            stdout0 = stdout1
-#            tme.sleep(.5)
+    #    import time as tme
+    #    vectorlen = np.vectorize(len)
+    #
+    #    stdout0 = np.array(async_res.stdout)
+    #    while not async_res.ready():
+    #        #ids = async_res.engine_id
+    #        stdout1 = np.array(async_res.stdout)
+    #        where = stdout1 != stdout0
+    #        if np.any(where):
+    #            oldlens = vectorlen(stdout0[where])
+    #            changedstdouts = ['Kernel '
+    #                              #+ str(ids[i]) + ':  '
+    #                              + str(np.arange(len(where))[where][i]) + ':  '
+    #                              + '\n'
+    #                              + val[oldlength:]
+    #                              for i, [val, oldlength] in enumerate(zip(stdout1[where], oldlens))]
+    #
+    #            sys.stdout.write('\r ' + ' '.join(changedstdouts))
+    #            sys.stdout.flush()
+    #            stdout0 = stdout1
+    #            tme.sleep(.5)
     pass
