+++ vsc_scripts/config_vsc.ini
